+++ patches/Accessibility/strings_fr.cpp
@@ -462,6 +462,8 @@ const char* Get(Id id) {
         case Id::FmtModSettingOption:         return "%s : %s";
         case Id::ModSettingCueVolume:         return "Volume des indices sonores";
         case Id::FmtModSettingSlider:         return "%s : %d pour cent";
+        case Id::ModSettingUrgentVolume:      return "Volume des annonces vocales";
+        case Id::ModSettingUrgentVolumePreview: return "Exemple d'annonce";
 
         case Id::ModSettingAudioGlossary:           return "Glossaire audio";
         case Id::ModSettingsAudioGlossaryOpened:    return "Glossaire audio ouvert";